requirements.txt
fastapi
uvicorn[standard]
jinja2
motor
httpx
pydantic-settings
bson